+++ lib/apitizer.rb
@@ -6,7 +6,6 @@ require_relative 'apitizer/helper'
 
 require_relative 'apitizer/routing'
 require_relative 'apitizer/connection'
-require_relative 'apitizer/processing'
 
 require_relative 'apitizer/result'
 require_relative 'apitizer/base'