+++ lib/apitizer/connection/format/yaml.rb
@@ -1,9 +1,13 @@
 require 'yaml'
 
 module Apitizer
-  module Processing
-    module Parser
+  module Connection
+    module Format
       class YAML
+        def mime_type
+          'application/x-yaml'
+        end
+
         def process(data)
           ::YAML.load(data)
         rescue