+++ lib/apitizer/routing/map.rb
@@ -1,6 +1,6 @@
 module Apitizer
   module Routing
-    class Mapper
+    class Map
       extend Forwardable
 
       def_delegator :@root, :define_address
@@ -11,20 +11,13 @@ module Apitizer
       end
 
       def trace(action, *arguments)
-        path = @root.trace(*arguments)
-        raise Error, 'Not permitted' unless path.permitted?(action)
+        path = @root.trace(*arguments) or raise Error, 'Not found'
+        raise Error, 'Not permitted' unless path.permit?(action)
         path
       end
 
       def define(&block)
         proxy = Proxy.new(self)
-        proxy.instance_eval(&block)
-      end
-
-      def define_scope(path, parent: @root, &block)
-        child = Node::Scope.new(path)
-        parent.append(child)
-        proxy = Proxy.new(self, parent: child)
         proxy.instance_eval(&block)
       end