+++ lib/apitizer/routing/node/root.rb
@@ -2,12 +2,18 @@ module Apitizer
   module Routing
     module Node
       class Root < Base
+        def recognize?(steps)
+          true
+        end
+
         def define_address(address, *_)
           @address = address
         end
 
-        def process(path, steps)
-          path << @address unless @address.nil?
+        private
+
+        def walk(steps, path)
+          path.advance(@address, node: self) if @address
         end
       end
     end