+++ spec/apitizer/connection/dispatcher_spec.rb
@@ -1,37 +1,35 @@
 require 'spec_helper'
 
-describe Apitizer::Connection::Dispatcher do
+RSpec.describe Apitizer::Connection::Dispatcher do
   extend ResourceHelper
   include ResourceHelper
 
-  let(:headers) { { 'Secret-Token' => 'arbitrary' } }
-  let(:address) { 'https://service.com/api/v1/json/articles' }
-  let(:subject) do
-    Apitizer::Connection::Dispatcher.new(
-      dictionary: rest_http_dictionary, headers: headers)
-  end
+  let(:address) { 'https://service.com/api/articles' }
 
-  def create_request(action, address)
-    double(action: action, address: address, parameters: {})
+  def create_request(method)
+    double(method: method, address: address, parameters: {})
   end
 
   describe '#process' do
-    restful_actions.each do |action|
-      method = rest_http_dictionary[action]
+    { :json => '{}', :yaml => '---' }.each do |format, sample|
+      context "when interacting in #{ format }" do
+        let(:subject) { Apitizer::Connection::Dispatcher.new(format: format) }
 
-      context "when sending #{ action } Requests" do
-        it 'sets the token header' do
-          stub = stub_http_request(method, address)
-          response = subject.process(create_request(action, address))
-          expect(stub).to \
-            have_requested(method, address).with(headers: headers)
-        end
+        http_methods.each do |method|
+          context "when performing #{ method } operations" do
+            it 'uses propoer HTTP methods' do
+              stub = stub_http_request(method, address).to_return(body: sample)
+              response = subject.process(create_request(method))
+              expect(stub).to have_been_requested
+            end
 
-        it 'returns Responses' do
-          stub_http_request(method, address).
-            to_return(code: '200', body: 'Hej!')
-          response = subject.process(create_request(action, address))
-          expect([ response.code, response.body ]).to eq([ 200, 'Hej!' ])
+            it 'sets proper headers' do
+              stub = stub_http_request(method, address).to_return(body: sample).
+                with(headers: { 'Accept' => mime_type_dictionary[format] })
+              response = subject.process(create_request(method))
+              expect(stub).to have_been_requested
+            end
+          end
         end
       end
     end