+++ spec/spec_helper.rb
@@ -4,8 +4,5 @@ require 'webmock/rspec'
 require 'apitizer'
 
 RSpec.configure do |config|
-  config.treat_symbols_as_metadata_keys_with_true_values = true
-  config.run_all_when_everything_filtered = true
-  config.filter_run :focus
-  config.order = 'random'
+  config.disable_monkey_patching!
 end