+++ spec/support/factory_helper.rb
@@ -1,23 +1,43 @@
 module FactoryHelper
+  include ResourceHelper
+
   def create_tree(*names)
     operations = names.last.is_a?(Hash) ? names.pop : {}
-    root = Apitizer::Routing::Node::Root.new
+
+    root = build_root
+
     leaf = names.inject(root) do |parent, object|
       if object.is_a?(Array)
         name, only = *object
-        node = Apitizer::Routing::Node::Collection.new(name, only: only)
+        node = build_collection(name, only: only)
       else
         name = object
-        node = Apitizer::Routing::Node::Collection.new(name)
+        node = build_collection(name)
       end
       parent.append(node)
       node
     end
+
     operations.each do |name, action|
-      operation = Apitizer::Routing::Node::Operation.new(
-        name, action: action, on: :member)
+      on = restful_member_actions.include?(action) ? :member : :collection
+      operation = build_operation(name, action: action, on: on)
       leaf.append(operation)
     end
+
     root
+  end
+
+  private
+
+  def build_root(*arguments)
+    Apitizer::Routing::Node::Root.new(*arguments)
+  end
+
+  def build_collection(*arguments)
+    Apitizer::Routing::Node::Collection.new(*arguments)
+  end
+
+  def build_operation(*arguments)
+    Apitizer::Routing::Node::Operation.new(*arguments)
   end
 end