+++ spec/support/resource_helper.rb
@@ -15,4 +15,12 @@ module ResourceHelper
     { :index => :get, :show => :get, :create => :post,
       :update => :put, :delete => :delete }
   end
+
+  def http_methods
+    [ :get, :post, :put, :patch, :delete ]
+  end
+
+  def mime_type_dictionary
+    { :json => 'application/json', :yaml => 'application/x-yaml' }
+  end
 end