+++ CHANGELOG.md
@@ -1,1 +1,8 @@
+## Apitizer (master)
+* A new strategy for searching request endpoints.
+* Support for `on: :collection` in addition to `on: :member`.
+* Support for `except: [ ... ]` in addition to `only: [ ... ]`.
+* Setting the Accept header according to the data format.
+* Sending the parameter of a POST/PUT/PATCH/DELETE request in its body.
+
 ## Apitizer 0.0.1 (June 1, 2014)